+ Projects WP8/W8 for Direct3D rendering - after these mods it should render red solid background on both platforms

+++ cocos/2d/CCLabel.h
@@ -273,7 +273,7 @@ public:
     CC_DEPRECATED_ATTRIBUTE virtual void setFontDefinition(const FontDefinition& textDefinition);
     CC_DEPRECATED_ATTRIBUTE const FontDefinition& getFontDefinition() const { return _fontDefinition; }
 
-    CC_DEPRECATED_ATTRIBUTE int getCommonLineHeight() const { return getLineHeight();}
+    CC_DEPRECATED_ATTRIBUTE float getCommonLineHeight() const { return getLineHeight();}
 
 protected:
     void onDraw(const Mat4& transform, bool transformUpdated);
